_all_agents/_agent_mongo/llm_chat.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv
import os

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

llm = ChatOpenAI(
    model="gpt-4o-mini",
    openai_api_key=OPENAI_API_KEY,
    temperature=0.7,
)

async def get_response(message):
    response = await llm.ainvoke({"messages": [{"role": "user", "content": message}]})
    return response.content