requirements.txt
langchain
deepagents
langgraph
numpy
spacy
nltk
fastapi
uvicorn
redis
celery
motor
pymongo
qdrant_client
fastembed
pydantic
python-dotenv
langchain-community
langchain-core
langchain-openai
langchain-deepseek
langchain-ollama
bcrypt==4.0.1
python-jose
passlib
langgraph-checkpoint-mongodb
fastmcp
langchain_mcp_adapters